Revert commit for 'New About screen'
