+++ app/screens/about/title.tsx
@@ -2,34 +2,26 @@
 // See LICENSE.txt for license information.
 
 import React from 'react';
-import {Text} from 'react-native';
 
 import FormattedText from '@components/formatted_text';
 import {useTheme} from '@context/theme';
 import {t} from '@i18n';
 import {makeStyleSheetFromTheme} from '@utils/theme';
-import {typography} from '@utils/typography';
-
-const getStyleSheet = makeStyleSheetFromTheme((theme) => {
-    return {
-        title: {
-            ...typography('Heading', 800, 'SemiBold'),
-            color: theme.centerChannelColor,
-            paddingHorizontal: 36,
-        },
-        spacerTop: {
-            marginTop: 8,
-        },
-        spacerBottom: {
-            marginBottom: 8,
-        },
-    };
-});
 
 type TitleProps = {
     config: ClientConfig;
     license: ClientLicense;
 }
+
+const getStyleSheet = makeStyleSheetFromTheme((theme) => {
+    return {
+        title: {
+            fontSize: 22,
+            color: theme.centerChannelColor,
+        },
+    };
+});
+
 const Title = ({config, license}: TitleProps) => {
     const theme = useTheme();
     const style = getStyleSheet(theme);
@@ -46,23 +38,13 @@ const Title = ({config, license}: TitleProps) => {
             defaultMessage = 'Enterprise Edition';
         }
     }
-
     return (
-        <>
-            <Text
-                style={[style.title, style.spacerTop]}
-                testID='about.site_name'
-            >
-                {`${config.SiteName} `}
-            </Text>
-            <FormattedText
-                id={id}
-                defaultMessage={defaultMessage}
-                style={[style.title, style.spacerBottom]}
-                testID='about.title'
-            />
-        </>
-
+        <FormattedText
+            id={id}
+            defaultMessage={defaultMessage}
+            style={style.title}
+            testID='about.title'
+        />
     );
 };
