--- app/screens/settings/about/index.tsx
@@ -1,18 +1,0 @@
-// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-
-import {withDatabase} from '@nozbe/watermelondb/DatabaseProvider';
-import withObservables from '@nozbe/with-observables';
-
-import {observeConfig, observeLicense} from '@queries/servers/system';
-
-import About from './about';
-
-import type {WithDatabaseArgs} from '@typings/database/database';
-
-const enhanced = withObservables([], ({database}: WithDatabaseArgs) => ({
-    config: observeConfig(database),
-    license: observeLicense(database),
-}));
-
-export default withDatabase(enhanced(About));
